RAG/requirements.txt
streamlit
openai
faiss-cpu
langchain
langchain-community
PyMuPDF
python-dotenv
pdfplumber
tabula-py
tiktoken
amazon-textract-caller
amazon-textract-textractor
amazon-textract-caller
pdf2image
pytesseract
mistralai